backend/requirements.txt
flask
flask-cors
torch
torchvision
pillow
numpy
tensorflow
efficientnet-pytorch
opencv-python
scikit-learn
python-dotenv
gunicorn